Fixed issues that came up while testing
Updated css so phone display is cleaner
Moved "Total meals Leftover meals" to the left a bit.
Moved the site phone number to the start and added to pdf
Fixed the meal temperature not displaying on the pdf.
Removed extraneous meal temperature at the end of the process -
Fixed the Additional meals not displaying on the pdf.
Changed so Additional meals only displays when needed
Replace the jsignature file with older version to fix signature issues of the png signature causes the PDFs do not open using Adobe Acrobat and the jpeg option produces black background.
The signature input box is sometimes created incorrectly and won't allow an input without resizing the screen.
Removed most occurences of col-md-6 as it disallows use of full width of phone
Added meal check because on Android when you tap through the input boxes on the first screen and then select Go on the last one, it will take you into the second screen without selecting a meal. This is not a problem with Apple.